refactor(delta-scan): split delta_scan.rs into model + scan modules (#1124)
Behavior-preserving module split of the ~3,108-line
reader/delta_scan.rs into a module directory:

- delta_scan/mod.rs   - entry point + pub use re-exports (public
  surface unchanged: DeltaRecord, CellDelta, CellMeta, RangeBound,
  RowKeys, ScanSummary, ScanSummaryHandle, scan_delta, ScanDeltaOutput)
- delta_scan/model.rs - data-model types + their unit tests
- delta_scan/scan.rs  - scan_delta streaming driver, find_data_db,
  and the integration/parser test suite

Tests follow their code; all 48 delta_scan tests preserved and
passing. No semantic changes. Part of epic #1116.
